ultimas modificaciones

+++ Universidad88/src/entidades/Inscripcion.java
@@ -5,19 +5,19 @@ public class Inscripcion {
     private int idInscripcion;
     private Alumno alumno; 
     private Materia materia;
-    private int nota; 
+    private double nota; 
 
     public Inscripcion() {
     }
 
-    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, int nota) {
+    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, double nota) {
         this.idInscripcion = idInscripcion;
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
     }
 
-    public Inscripcion(Alumno alumno, Materia materia, int nota) {
+    public Inscripcion(Alumno alumno, Materia materia, double nota) {
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
@@ -47,11 +47,11 @@ public class Inscripcion {
         this.materia = materia;
     }
 
-    public int getNota() {
+    public double getNota() {
         return nota;
     }
 
-    public void setNota(int nota) {
+    public void setNota(double nota) {
         this.nota = nota;
     }
